Rename 'link' to 'linkList'
