+++ docs/examples/other/linkList/LinkBasic.txt
