+++ src/controls/index.ts
@@ -12,7 +12,7 @@ export { default as VdToggle } from 'src/controls/form/toggle/VdToggle.vue';
 export { default as VdArticle } from 'src/controls/layout/common/VdArticle.vue';
 export { default as VdMain } from 'src/controls/layout/common/VdMain.vue';
 export { default as VdFlexbox } from 'src/controls/layout/flexbox/VdFlexbox.vue';
-export { default as VdLinkList } from 'src/controls/link/VdLinkList.vue';
+export { default as VdLinkList } from 'src/controls/linkList/VdLinkList.vue';
 export { default as VdSubnav } from 'src/controls/navigation/subnav/VdSubnav.vue';
 export { default as VdSubnavItem } from 'src/controls/navigation/subnav/VdSubnavItem.vue';
 export { default as VdTabs } from 'src/controls/navigation/tabs/VdTabs.vue';
