+++ src/index.scss
@@ -16,7 +16,7 @@
 @import 'controls/layout/common/common';
 @import 'controls/layout/common/main';
 @import 'controls/layout/flexbox/flexbox';
-@import 'controls/link/linklist';
+@import 'controls/linkList/linklist';
 @import 'controls/navigation/subnav/subnav';
 @import 'controls/navigation/tabs/tabs';
 @import 'controls/note/note';
